rag_chatbot/requirements.txt
google-genai>=1.0.0
numpy>=1.23.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
